move DisconnectReq to agent from aclk
