+++ Makefile.am
@@ -665,8 +665,6 @@ ACLK_NG_FILES += \
     aclk/schema-wrappers/alarm_stream.h \
     aclk/schema-wrappers/alarm_config.cc \
     aclk/schema-wrappers/alarm_config.h \
-    aclk/schema-wrappers/cloud_cmds.cc \
-    aclk/schema-wrappers/cloud_cmds.h \
     aclk/schema-wrappers/node_info.cc \
     aclk/schema-wrappers/node_info.h \
     aclk/schema-wrappers/schema_wrappers.h \
@@ -676,7 +674,7 @@ ACLK_NG_FILES += \
 
 ACLK_NG_PROTO_DEFINITIONS = \
     aclk/aclk-schemas/proto/aclk/v1/lib.proto \
-    aclk/aclk-schemas/proto/aclk/v1/cmd.proto \
+    aclk/aclk-schemas/proto/agent/v1/disconnect.proto \
     aclk/aclk-schemas/proto/agent/v1/connection.proto \
     aclk/aclk-schemas/proto/alarm/v1/config.proto \
     aclk/aclk-schemas/proto/alarm/v1/stream.proto \
@@ -707,8 +705,8 @@ ACLK_NG_PROTO_BUILT_FILES = aclk/aclk-schemas/proto/agent/v1/connection.pb.cc \
     aclk/aclk-schemas/proto/chart/v1/config.pb.h \
     aclk/aclk-schemas/proto/aclk/v1/lib.pb.cc \
     aclk/aclk-schemas/proto/aclk/v1/lib.pb.h \
-    aclk/aclk-schemas/proto/aclk/v1/cmd.pb.cc \
-    aclk/aclk-schemas/proto/aclk/v1/cmd.pb.h \
+    aclk/aclk-schemas/proto/agent/v1/disconnect.pb.cc \
+    aclk/aclk-schemas/proto/agent/v1/disconnect.pb.h \
     aclk/aclk-schemas/proto/alarm/v1/config.pb.cc \
     aclk/aclk-schemas/proto/alarm/v1/config.pb.h \
     aclk/aclk-schemas/proto/alarm/v1/stream.pb.cc \
@@ -753,8 +751,8 @@ aclk/aclk-schemas/proto/aclk/v1/lib.pb.cc \
 aclk/aclk-schemas/proto/aclk/v1/lib.pb.h: aclk/aclk-schemas/proto/aclk/v1/lib.proto
 	$(PROTOC) -I=aclk/aclk-schemas --cpp_out=$(builddir)/aclk/aclk-schemas $^
 
-aclk/aclk-schemas/proto/aclk/v1/cmd.pb.cc \
-aclk/aclk-schemas/proto/aclk/v1/cmd.pb.h: aclk/aclk-schemas/proto/aclk/v1/cmd.proto
+aclk/aclk-schemas/proto/agent/v1/disconnect.pb.cc \
+aclk/aclk-schemas/proto/agent/v1/disconnect.pb.h: aclk/aclk-schemas/proto/agent/v1/disconnect.proto
 	$(PROTOC) -I=aclk/aclk-schemas --cpp_out=$(builddir)/aclk/aclk-schemas $^
 
 aclk/aclk-schemas/proto/alarm/v1/config.pb.cc \
