+++ aclk/schema-wrappers/connection.h
@@ -27,6 +27,15 @@ typedef struct {
 
 char *generate_update_agent_connection(size_t *len, const update_agent_connection_t *data);
 
+struct disconnect_cmd {
+    uint64_t reconnect_after_s;
+    int permaban;
+    uint32_t error_code;
+    char *error_description;
+};
+
+struct disconnect_cmd *parse_disconnect_cmd(const char *data, size_t len);
+
 #ifdef __cplusplus
 }
 #endif
